+++ geosoft.c
@@ -195,7 +195,7 @@ char *filename;
                    WSF_NO_MENUBAR | WSF_CLOSE | WSF_SIZE |
                    WSF_INVISIBLE | WSF_ICONIZABLE,
 #endif
-                   EM_ALL, anomImageEventHandler, (long) anomImageData)))
+                   EM_ALL, anomImageEventHandler, (intptr_t) anomImageData)))
    {
       finishLongJob ();
       xvt_dm_post_error("Cannot Create Image Display Window.");
@@ -245,7 +245,7 @@ char *filename;
                                /* rect is the size of the data array */
    updateAnomaliesImageDisplay(imageWindow);
    
-   xvt_vobj_set_data (imageWindow, (long) anomImageData);
+   xvt_vobj_set_data (imageWindow, (intptr_t) anomImageData);
 
    finishLongJob ();
